working strings MR generation to android & ios
