+++ sample/mpp-library/src/commonMain/resources/MR/base/strings.xml
@@ -3,4 +3,5 @@
     <string name="test">test</string>
     <string name="test2">test 2</string>
     <string name="test3">test 3</string>
+    <string name="common.name">Test Project</string>
 </resources>
