+++ sample/mpp-library/src/commonMain/resources/MR/ru/strings.xml
@@ -3,4 +3,5 @@
     <string name="test">тест</string>
     <string name="test2">тест 2</string>
     <string name="test3">тест 3</string>
+    <string name="common.name">Тестовый проект</string>
 </resources>
